Move EnergySetter to apple2
