+++ src/dodal/devices/apple2_undulator.py
@@ -10,6 +10,7 @@ from ophyd_async.core import (
     AsyncStatus,
     Device,
     FlyMotorInfo,
+    Reference,
     SignalR,
     SignalW,
     StandardReadable,
@@ -20,12 +21,14 @@ from ophyd_async.core import (
     derived_signal_rw,
     observe_value,
     soft_signal_r_and_setter,
+    soft_signal_rw,
     wait_for_value,
 )
 from ophyd_async.epics.core import epics_signal_r, epics_signal_rw, epics_signal_w
 from ophyd_async.epics.motor import Motor
 from pydantic import BaseModel, ConfigDict, RootModel
 
+from dodal.devices.pgm import PGM
 from dodal.log import LOGGER
 
 T = TypeVar("T")
@@ -796,3 +799,44 @@ class Apple2(abc.ABC, StandardReadable, Movable, Preparable, Flyable):
 
         LOGGER.warning("Unable to determine polarisation. Defaulting to NONE.")
         return Pol.NONE, 0.0
+
+
+ID_TYPE = TypeVar("ID_TYPE", bound=Apple2)
+
+
+class EnergySetter(abc.ABC, StandardReadable, Movable[float], Generic[ID_TYPE]):
+    """
+    Compound device to set both ID and PGM energy at the same time.
+
+    """
+
+    def __init__(self, id: ID_TYPE, pgm: PGM, name: str = "") -> None:
+        """
+        Parameters
+        ----------
+        id:
+            An Apple2 device (should be I10Apple2 or subclass).
+        pgm:
+            A PGM/mono device.
+        name:
+            New device name.
+        """
+
+        self.id = id
+        self.pgm_ref = Reference(pgm)
+        self.add_readables(
+            [self.id.energy, self.pgm_ref().energy.user_readback],
+            StandardReadableFormat.HINTED_SIGNAL,
+        )
+
+        with self.add_children_as_readables(StandardReadableFormat.CONFIG_SIGNAL):
+            self.energy_offset = soft_signal_rw(float, initial_value=0)
+        super().__init__(name=name)
+
+    @AsyncStatus.wrap
+    async def set(self, value: float) -> None:
+        LOGGER.info(f"Moving f{self.name} energy to {value}.")
+        await asyncio.gather(
+            self.id.set(value=value + await self.energy_offset.get_value()),
+            self.pgm_ref().energy.set(value),
+        )
